requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.2
django-allauth==0.51.0
django-autoslug==1.9.8
django-celery-beat==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-countries==7.4.2
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-timezone-field==5.0
Django-Verify-Email==2.0.3
django-webpush==0.3.4
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hiredis==2.0.0
http-ece==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonfield==3.1.0
kombu==5.2.4
msgpack==1.0.4
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
progressbar2==4.2.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
python-utils==3.4.5
python3-openid==3.2.0
pytz==2022.6
pywebpush==1.9.4
razorpay==1.3.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tqdm==4.64.1
Twisted==22.10.0
txaio==22.2.1
types-cryptography==3.3.23
typing_extensions==4.4.0
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.5.1
